refactor(client): decompose BookmarkPicker into hook and subcomponents
Extracts useBookmarkPicker (debounced search, add/remove/section
mutations, paste-a-URL resolve-or-create), the BookmarkSearchDropdown
results panel, and BookmarkChip into their own modules, leaving
BookmarkPicker as composition. Stories pin the dropdown's five states
and the chip with/without sections.
